requirements.txt
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
Flask-MySQL==1.5.2
Flask-WTF==0.15.1
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
numpy==1.21.0
pandas==1.3.0
protobuf==3.17.3
pycodestyle==2.7.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.4.1
